chore(dependabot.yml): add grouping for GitHub Actions and pip dependencies
Add groups for GitHub Actions and pip dependencies in the Dependabot
configuration to allow for more organized and manageable updates. This
ensures that updates are grouped together, making it easier to review
and merge changes.

style(action.yml): standardize quotes and improve comment consistency

Standardize the use of double quotes in the setup/action.yml file for
consistency. Adjust comments to use consistent phrasing, improving
readability and maintaining a uniform style throughout the file.